Unify Space and Convo Cards into generic RoomCard
