+++ app/lib/features/space/pages/chats_page.dart
@@ -3,7 +3,7 @@ import 'package:acter/common/providers/space_providers.dart';
 import 'package:acter/common/toolkit/buttons/inline_text_button.dart';
 import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
 import 'package:acter/common/utils/routes.dart';
-import 'package:acter/common/widgets/chat/loading_convo_card.dart';
+import 'package:acter/features/space/widgets/loading_convo_card.dart';
 import 'package:acter/common/widgets/empty_state_widget.dart';
 import 'package:acter/features/space/widgets/related/chats_helpers.dart';
 import 'package:atlas_icons/atlas_icons.dart';
